pyproject.toml
[tool.poetry]
name = "clinical-mas-planner"
version = "0.1.0"
description = "Multi-Agent Diagnostic Reasoning System for Clinical Decision Support"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
typer = {extras = ["all"], version = "^0.9.0"}
openai = "^1.10.0"
anthropic = "^0.18.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
httpx = "^0.26.0"
sqlmodel = "^0.0.14"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
mypy = "^1.8.0"
black = "^24.1.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
mas = "src.app:cli_app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
